+++ modules/graphical-minimal.nix
@@ -1,6 +1,21 @@
 {
   imports = [ "firefox" "fonts" ];
 
+  darwinModule = { pkgs, ... }: {
+    environment.systemPackages =
+      builtins.attrValues { inherit (pkgs) rectangle; };
+
+    # Close Terminal if shell exited cleanly
+    system.activationScripts.extraUserActivation.text = ''
+      plutil -replace "Window Settings.Basic.shellExitAction" -integer 1 ~/Library/Preferences/com.apple.Terminal.plist
+    '';
+
+    # WORKAROUND: Screensaver starts on the login screen and cannot be closed from VNC
+    system.activationScripts.extraActivation.text = ''
+      defaults write /Library/Preferences/com.apple.screensaver loginWindowIdleTime 0
+    '';
+  };
+
   nixosModule = { user, pkgs, ... }: {
     environment.systemPackages =
       builtins.attrValues { inherit (pkgs) pavucontrol; };