+++ modules/vscode.nix
@@ -11,20 +11,20 @@
       home.file.".vscode-server/extensions".source =
         config.home.file.".vscode/extensions".source;
 
-      # WORKAROUND: home-manager uses `sudo -u` to run activation scripts as the correct user
-      #             however dockutil uses `SUDO_USER` to determine the user who ran the command
-      #             meaning that it attempts to edit root's Dock intead of the current user
-      home.activation.addVSCodeToDock = lib.mkIf hostPlatform.isDarwin
-        (lib.hm.dag.entryAfter [ "writeBoundary" ] ''
-          echo "adding Visual Studio Code.app to the dock"
-          SUDO_USER= ${pkgs.dockutil}/bin/dockutil --add "${config.programs.vscode.package}/Applications/Visual Studio Code.app" --replacing "Visual Studio Code"
-        '');
+      home.file.".vscode-server/data/Machine/settings.json".source =
+        config.home.file."${
+          if hostPlatform.isDarwin then
+            "Library/Application Support"
+          else
+            config.xdg.configHome
+        }/Code/User/settings.json".source;
 
       programs.vscode.enable = true;
       # Don't install `vscode` unless `graphical` module is specified
       programs.vscode.package = lib.mkDefault (pkgs.emptyDirectory // {
         pname = "vscode";
-        version = "";
+        # Required for version check to generate extensions.json
+        version = "1.74.0";
       });
       programs.vscode.extensions = [
         pkgs.vscode-extensions.asvetliakov.vscode-neovim